refactor : MVC 패턴으로 리펙터링. 스켈레톤만 짜 놓음
