+++ src/main/java/baseball/BaseballApplication.java
@@ -1,0 +1,4 @@
+package baseball;
+
+public class BaseballApplication {
+}
