+++ src/main/java/baseball/view/OutputView.java
@@ -1,0 +1,4 @@
+package baseball.view;
+
+public class OutputView {
+}
